New functionality and name changes
Renamed bus_routes.js to pat.js, with "pat" short for "Port Authority
Transit." Adjusted folder structure to improve organization.
Added functionality to pat.js:
get_stops accepts a route, direction, and rtpi and returns an array of stops.
get_vehicles accepts a route and returns a list of vehicles currently
operating on that route.
get_dir accepts an array of routes and adds another property to each
element of that array specifying the possible directions. For Pittsburgh
it's trivial; there's only INBOUND and OUTBOUND.
get_routes accepts, but does not require, a search term, searches all
available routes, and returns an array of routes with matching names. If search term
is an empty string get_routes returns an array of all known routes.