Added new function, rewrote others to get trips.
Added new function routeIdFromTripId() that takes a string tripId and
searches the csv file trips.txt for a matching trip. If found, it
returns the route and headsign of that trip. Rewrote nearestTrips().
Runs effectively, but slowly. Need to see if there's a faster way to
parse. Maybe by sorting once and keeping relevant data in a new file.